Fix dev server's handling of non-script assets
Because the app is written in TypeScript, which needs to be compiled,
the server script that is actually run is in the `dist/` directory
rather than in `src/`.  This means that our build process must include
copying other files (templates, stylesheets, etc.) into the `dist/`
directory as well.

Our current solution, when running `npm run start:dev`, results in
initially copying the assets and compiling the TypeScript files, then
only rebuilding when a .ts, .js, or .json file changes.  The result is
that changes to an HTML file, for example, won't be reflected until
some ts/js/json file is changed or `nodemon` itself is restarted
(generally `docker-compose stop && docker-compose up`).

This commit adjusts the npm script so that it starts up 2 simultaneous
instances of `nodemon` -- one for the scripts and one for the rest of
the assets.

A different solution that would have been conceptually simpler (and run
fewer processes) would be to have a single `nodemon` instance watch
everything and copy/rebuild on any change.  However, it takes
substantially longer to rebuild and restart the server than to copy
files, so I thought it would be better to separate the tasks in order to
speed up the turnaround time when only working on the frontend content.